requirements.txt
streamlit==1.32.0
openai-whisper==20231117
python-docx==1.1.0
numpy==1.24.3
soundfile==0.12.1
librosa==0.10.1
torch>=2.0.0
tqdm>=4.42.1
more-itertools>=8.0.0
transformers>=4.19.0
ffmpeg-python>=0.2.0